Streamlit/requirements.txt
numpy>=1.14.5
opencv_python==4.5.5.64
Pillow==9.1.0
streamlit==1.3.1
tensorflow_cpu==2.8.0
